Issue3 RC 1
- Extracted <ApiInteractionInterface />
